fix: 修复3个类型不匹配编译错误
1. MarkdownRenderer: Int * Dp → Dp * Int
2. PersonalWakeListener: Float / Int 类型推断 → 显式 toFloat()
3. VoiceWakeService: readCount 类型推断 → 显式声明 Int

+++ AndroidDevAgent/app/src/main/java/com/example/androiddevagent/voice/PersonalWakeListener.kt
@@ -122,7 +122,7 @@ class PersonalWakeListener(
         }
 
         // 归一化距离
-        val normalizedDistance = minDistance / (features.size + templates.firstOrNull()?.size?.let { it / 2 } ?: 1)
+        val normalizedDistance = minDistance / (features.size.toFloat() + (templates.firstOrNull()?.size?.let { it / 2 } ?: 1).toFloat())
 
         return normalizedDistance <= dtwThreshold
     }
